Add Electrum peers discovery
Checker can fetch discover peers of each server and then continuously
add new nodes that it can connect to. There should be additional checks
in place that can assess the health of a connected node, and a potential
retry if a node is no-longer healthy.

- Updated go-electrum library to a different fork, so that connection
timeouts are enforced by node.ConnectTCP and node.ConnectSSL