Handle closed channels in accounter logic
Some backends might close their channels and accounter should know how
to deal with that. When a closed channel is bein read, a zero value is returned
immediately and when a nil channel is read it blocks forever. Select
skips the channels that are blocking and hence nil channels are an
efficient way to ignore channels that are done.

Backend shutdown got simplified and doesn't require explicit calls to
`Dec()` to subtract from the WaitGroup. In fact, WaitGroup got removed
from backends.

Also, addresses race conditions in the reporter where reads and writes
to reporter's fields were not synchronized. Now adding, getting and
setting a value is atomic.